Move stuff around and start to play with fancier gen
I have to set additional response codes and its being a real pain
