+++ app/Main.hs
@@ -1,28 +1,17 @@
-{-# LANGUAGE OverloadedLists #-}
 import Control.Monad.Reader (runReaderT)
-import Data.Swagger
 import Data.Proxy (Proxy(..))
 import Database.PostgreSQL.Simple (connectPostgreSQL)
 import Network.Wai (Application)
 import Network.Wai.Handler.Warp (run)
 import Network.Wai.Middleware.RequestLogger (logStdoutDev)
 import Servant (Handler, Server, (:<|>)(..), (:~>)(NT), enter, serve)
-import Servant.Swagger
 import System.Environment (lookupEnv)
 import Text.Read (readMaybe)
 
 import Config (Config(..), ConfigM(..))
-import Routes (API)
+import API (API)
 import Routes.Video (videoServer)
 import Routes.VideoLibrary (videoLibraryServer)
-
-import Types.Video
-import Types.VideoLibrary
-import Types.VideoFile
-import Data.Aeson.Encode
-import qualified Data.ByteString.Lazy.Char8 as BL8
-import Control.Lens
-import Data.Text (Text)
 
 maybeGetPort :: IO (Maybe Int)
 maybeGetPort = do
@@ -55,28 +44,8 @@ makeHandler config = NT (\m -> runReaderT (runConfigM m) config)
 app :: Config -> Application
 app config = serve api (server config)
 
-myOptions = defaultSchemaOptions{unwrapUnaryRecords=True}
-
-instance ToSchema VideoFile
-instance ToSchema VideoFileId where declareNamedSchema = genericDeclareNamedSchema myOptions
-instance ToSchema (Video VideoId)
-instance ToSchema (Video NoId) where
-  declareNamedSchema _ = do
-    nameSchema <- declareSchemaRef (Proxy :: Proxy Text)
-    return $ NamedSchema (Just "NewVideo") $ mempty
-      & type_ .~ SwaggerObject
-      & properties .~ [ ("name", nameSchema) ]
-      & required .~ [ "name" ]
-instance ToSchema VideoId
-instance ToSchema VideoLibrary
-instance ToSchema VideoLibraryId where declareNamedSchema = genericDeclareNamedSchema myOptions
-instance ToSchema VideoFileStorageId where declareNamedSchema = genericDeclareNamedSchema myOptions
-instance ToParamSchema VideoId
-instance ToParamSchema VideoLibraryId
-
 main :: IO ()
 main = do
-  BL8.writeFile "swagger.json" $ encode $ toSwagger (Proxy :: Proxy API)
   -- Connection info gets passed via environment variables
   config <- getConfig
   port <- getPort
