+++ swagger.json
@@ -1,0 +1,142 @@
+{
+    "swagger": "2.0",
+    "info": {
+        "version": "",
+        "title": ""
+    },
+    "definitions": {
+        "VideoLibraryId": {
+            "maximum": 9223372036854775807,
+            "minimum": -9223372036854775808,
+            "type": "integer"
+        },
+        "VideoLibrary": {
+            "required": [
+                "videoLibraryId",
+                "videoLibraryUrl"
+            ],
+            "type": "object",
+            "properties": {
+                "videoLibraryId": {
+                    "$ref": "#/definitions/VideoLibraryId"
+                },
+                "videoLibraryUrl": {
+                    "type": "string"
+                }
+            }
+        }
+    },
+    "paths": {
+        "/video_libraries/{videoLibraryId}": {
+            "get": {
+                "responses": {
+                    "404": {
+                        "description": "`videoLibraryId` not found"
+                    },
+                    "200": {
+                        "schema": {
+                            "$ref": "#/definitions/VideoLibrary"
+                        },
+                        "description": ""
+                    }
+                },
+                "produces": [
+                    "application/json;charset=utf-8"
+                ],
+                "parameters": [
+                    {
+                        "maximum": 9223372036854775807,
+                        "minimum": -9223372036854775808,
+                        "required": true,
+                        "in": "path",
+                        "name": "videoLibraryId",
+                        "type": "integer"
+                    }
+                ]
+            }
+        },
+        "/video_libraries/{videoLibraryId}/new_files": {
+            "get": {
+                "responses": {
+                    "404": {
+                        "description": "`videoLibraryId` not found"
+                    },
+                    "200": {
+                        "schema": {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        },
+                        "description": ""
+                    }
+                },
+                "produces": [
+                    "application/json;charset=utf-8"
+                ],
+                "parameters": [
+                    {
+                        "maximum": 9223372036854775807,
+                        "minimum": -9223372036854775808,
+                        "required": true,
+                        "in": "path",
+                        "name": "videoLibraryId",
+                        "type": "integer"
+                    }
+                ]
+            }
+        },
+        "/video_libraries/{videoLibraryId}/all_files": {
+            "get": {
+                "responses": {
+                    "404": {
+                        "description": "`videoLibraryId` not found"
+                    },
+                    "422": {
+                        "description": "HGurra"
+                    },
+                    "200": {
+                        "schema": {
+                            "items": {
+                                "type": "string"
+                            },
+                            "type": "array"
+                        },
+                        "description": ""
+                    }
+                },
+                "produces": [
+                    "application/json;charset=utf-8"
+                ],
+                "parameters": [
+                    {
+                        "maximum": 9223372036854775807,
+                        "minimum": -9223372036854775808,
+                        "required": true,
+                        "in": "path",
+                        "name": "videoLibraryId",
+                        "type": "integer"
+                    }
+                ]
+            }
+        },
+        "/video_libraries": {
+            "get": {
+                "responses": {
+                    "200": {
+                        "schema": {
+                            "items": {
+                                "$ref": "#/definitions/VideoLibrary"
+                            },
+                            "type": "array"
+                        },
+                        "description": ""
+                    }
+                },
+                "produces": [
+                    "application/json;charset=utf-8"
+                ]
+            }
+        }
+    }
+}
