+++ tap-video.cabal
@@ -14,7 +14,8 @@ cabal-version:       >=1.10
 library
   hs-source-dirs:      src
   exposed-modules:
-                       Database.Error
+                       API
+                     , Database.Error
                      , Database.Video
                      , Database.VideoFile
                      , Database.VideoLibrary
@@ -27,6 +28,7 @@ library
                      , base >= 4.7 && < 5
                      , bytestring
                      , postgresql-simple
+                     , servant
                      , text
   default-language:    Haskell2010
   extensions:
@@ -35,12 +37,31 @@ library
                      , RecordWildCards
                      , TemplateHaskell
 
+executable tap-swagger-gen
+  hs-source-dirs:      swagger-gen
+  main-is:             Main.hs
+  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
+  build-depends:
+                       aeson
+                     , aeson-pretty
+                     , bytestring
+                     , base
+                     , lens
+                     , servant
+                     , servant-swagger
+                     , swagger2
+                     , tap-video
+                     , text
+  default-language:    Haskell2010
+  default-extensions:
+                       OverloadedStrings
+  other-extensions:
+                       DataKinds
 executable tap-video-exe
   hs-source-dirs:      app
   main-is:             Main.hs
   other-modules:
                        Config
-                     , Routes
                      , Routes.Video
                      , Routes.VideoLibrary
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N
